04. Iron Girder - 27 August
You will receive input lines in one of the following formats:
{townName}:{time}->{passengersCount} 
If you receive the line above, Iron Girder has travelled to certain town for a certain amount of time with certain count of passengers. You need to keep track for each town. You have to save the fastest time Iron Girder reaches a town and the total count of passengers for each town.
{townName}:ambush->{passengersCount}
If you receive the line above, somewhere along the track to the current town Iron Girder was ambushed and the passengers can't reach there. If this happens you need to set the time record for this town to "0" and remove the current count of passengers from the total count. If it's the first time Iron Girder travels to this town then you simply ignore this line.
When you receive "Slide rule" you end the program and print data for each town in the following format:
"{townName} -> Time: {fastestTime} -> Passengers: {totalCountPassengers}" 
The output should be ordered by best time and then by town's name. If a town is with no recorded time (the time is equal to 0) or there are no passengers (count is equal or less than 0) you should not print it.
Input:
Quirm:12->258
Ankh-Morpork:ambush->200
Ankh-Morpork:3->143
Sto-Lat:4->80
Ankh-Morpork:4->143
Ankh-Morpork:ambush->143
Sto-Lat:3->20
Ankh-Morpork:5->17
Slide rule