04. Anonymous Cache 05 November 2017
The Anonymous are storing data on their dataservers about their activities. The CIA has higher the greatest hacker in the world – You. Your job is to extract their data and send it to the CIA. It won’t be an easy task, Get Ready!
You will receive several input lines in one of the following formats:
•	{dataSet}
•	{dataKey} -> {dataSize} | {dataSet}
The dataSet and dataKey are both strings. The dataSize is an integer. The dataSets hold dataKeys and their dataSizes.
If you receive only a dataSet you should add it. If you receive a dataKey and a dataSize, you should add them to the given dataSet.
And here’s where the fun begins. If you receive a dataKey and a dataSize, but the given dataSet does NOT exist, you should STORE those keys and values in a cache. When the corresponding dataSet is added, you should check if the cache holds any keys and values referenced to it, and you should add them to the dataSet.
You should end your program when you receive the command “thetinggoesskrra”.  At that point you should extract the dataSet from the data with the HIGHEST dataSize (SUM of all its dataSizes), and you should print it. 
NOTE: Elements in the cache, should be CONSIDERED NON-EXISTANT. You should NOT count them in the final output.
In case there are NO dataSets in the data, you should NOT do anything.
Input:
Users
BankAccounts
ADDB444 -> 23111 | BankAccounts
Students -> 2000 | Users
Workers -> 24233 | Users
thetinggoesskrra